// src/app/shared/models/tipo-deduccion.enum.ts
export enum TipoDeduccion {
    Honorarios_Medicos = 1,
    Gastos_Medicos_Incapacidad = 2,
    Gastos_Funerales = 3,
    Donativos = 4,
    Intereses_Reales = 5,
    Aportaciones_Voluntarias = 6,
    Seguros_Gastos_Medicos = 7,
    Transportacion_Escolar = 8,
    Depositos_Cuentas_Ahorro = 9,
    Colegios = 10
  }
  